% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dall_amico.R
\name{DallAmico}
\alias{DallAmico}
\title{Compute total water content, ice content, and liquid water content of
       soil following Dall'Amico et al. (2011).}
\usage{
DallAmico(alpha, n, theta.sat, theta.res, Tg, Tm, psi0)
}
\arguments{
\item{alpha}{van Genuchten alpha [mm-1]}

\item{n}{van Genuchten n [-]}

\item{theta.sat}{Saturated water content [m3/m3]}

\item{theta.res}{Residual water content [m3/m3]}

\item{Tg}{Ground temperature [K], can be a vector.}

\item{Tm}{Melting temperature of water at atmospheric pressure [K], 
usually 273.15 K.}

\item{psi0}{Soil matric potential [m]. Matric potential of 0  corresponds 
to saturated conditions, smaller values imply unsaturated 
conditions.}
}
\value{
Returns a data frame with three columns: theta.v (total water content, 
        relative to soil volume, as a function of psi0), theta.i (ice content,
        relative to soil volume), and theta.w (liquid water content, relative
        to soil volume. The data frame has as manu rows as Tg has elements.
}
\description{
Compute total water content, ice content, and liquid water 
             content of soil based on van Genuchten parameters, temperature 
             and soil matric potential.
}
\details{
Details are found in: Dall’Amico, M., Endrizzi, S., Gruber, S., & 
         Rigon, R. (2011). A robust and energy-conserving model of freezing 
         variably-saturated soil. The Cryosphere, 5(2), 469-484. 
         doi:10.5194/tc-5-469-2011

         Typical values of van Genuchten parameters are found in Table 2 of
         Gubler, S., Endrizzi, S., Gruber, S., & Purves, R. S. (2013). 
         Sensitivities and uncertainties of modeled ground temperatures in 
         mountain environments. Geoscientific Model Development, 6(4), 
         1319–1336. doi:10.5194/gmd-6-1319-2013
}
\examples{
DallAmico(0.001, 1.6, 0.5, 0.02, 272, 273.15, 0)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
