% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_capacity.R
\name{HeatCapacity}
\alias{HeatCapacity}
\title{Heat capacity}
\usage{
HeatCapacity(mat)
}
\arguments{
\item{mat}{Data frame with relevant ground properties. Needed: mat$ice (ice
content, volumetric fraction) and mat$liq (liquid water
content, volumetric fraction).}
}
\value{
Returns the input data frame with the column cj updated in [J m-3 K-1]
        If this column does not exist in the input it is created.
}
\description{
Apparent heat capacity of soil or rock
}
\details{
Computes the apparent heat capacity of soil or rock based on
         the heat capacity of the consituents and the apparent heat 
         capacity parameterized using the derivative of the liquit 
         water content (see funtion "Unfrozen").
}
\examples{
mat <- HeatCapacity(mat)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
