% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_discretize.R
\name{SoilDiscretize}
\alias{SoilDiscretize}
\title{Discretize soil vertically}
\usage{
SoilDiscretize(dzmin, zmax, base)
}
\arguments{
\item{dzmin}{Minimal z-spacing [m] ("how fine is the grid")}

\item{zmax}{Depth of lowermost node [m] center ("how large is the domain")}

\item{base}{Resolution reduction ("how strong is the grid corsened with depth"); 
base = 1 corresponds to equal spacing, base = 1.5 implies each 
successive layer with depth begin 1.5 times thicker than the one 
above.}
}
\value{
data frame with layer thickness (dz) and depth of layer centre (z)
        below the surface
}
\description{
Generate thickness and depth of soil layers for numerical model
}
\details{
Makes a soil layer spacing based on the simple model 
         dz = dzmin * base^(n) where n is the layer number.
}
\examples{
soil <- SoilDiscretize(0.1, 20, 1.25)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
