% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therm_cond.R
\name{TermCond}
\alias{TermCond}
\title{Termal conductivity of soil/rock parametrized from its constituents
       and of snow parameterizad from density or constituents.}
\usage{
TermCond(mat, layers.sno = c(0), type.sno = "CALONNE",
  type.gnd = "COSENZA")
}
\arguments{
\item{mat}{Data frame with relevant ground properties. Needed: mat$ice (ice
content, volumetric fraction) and mat$liq (liquid water
content, volumetric fraction).}

\item{layers.sno}{Index to identify snow layers within 'mat'.}

\item{type.sno}{Character string indicating the type of parameterization to be
            used for snow; the standard is CALONNE. Available are:

            CALONNE: Calonne, N., Flin, F., Morin, S., Lesaffre, B., du 
            Roscoat, S. R., & Geindreau, C. (2011). Numerical and experimental
            investigations of the effective thermal conductivity of snow. 
            Geophysical Research Letters, 38(23), doi:10.1029/2011GL049234  
            "YEN": Yen, Y.-C. (1981). Review of thermal properties of 
             snow, ice and sea ice (34 pages). Hanover, NH, USA.

            COSENZA: Cosenza, P., Guerin, R., & Tabbagh, A. (2003). 
            Relationship between thermal conductivity and water content of 
            soils using numerical modelling. European Journal of Soil 
            Science, 54(3), 581–588. doi:10.1046/j.1365-2389.2003.00539.x

            STURM: Sturm, M., J. Holmgren, M. König, and K. Morris (1997),
            The thermal conductivity of seasonal snow, Journal of 
            Glaciology, 43(143), 26–41.

            JORDAN: Jordan, R. E., Andreas, E. L., & Makshtas, A. P. 
            (1999). Heat budget of snow-covered sea ice at North Pole 4. 
            Journal of Geophysical Research, 104(C4), 7785. 
            doi:10.1029/1999JC900011

            WILLIAMS: Figure 4.11 in "The Frozen Earth: fundamentals of 
            geocryology" by P. Williams and M. Smith, 1989.}

\item{type.gnd}{Character string indicating the type of parameterization to be
            used for ground; the standard is "COSENZA". Available are:

            COSENZA: Cosenza, P., Guerin, R., & Tabbagh, A. (2003). 
            Relationship between thermal conductivity and water content of 
            soils using numerical modelling. European Journal of Soil 
            Science, 54(3), 581–588. doi:10.1046/j.1365-2389.2003.00539.x

            GEOMETRIC: Geometric mean, intermediate mixed conductivity model, 
            approximation of randomly oriented consituent elements. 

            ARITHMETIC: Arithmetic mean, high mixed conductivity model, 
            approximation of consituent elements layered parallel to 
            temperature gradient.

            HARMONIC: Harmonic mean, low mixed conductivity model, 
            approximation of consituent elements layered normal to 
            temperature gradient.}
}
\value{
Returns the input data frame with the column kj updated in [W m-1 K-1]
        If this column does not exist in the input it is created.
}
\description{
Computes an estimated thermal condictivity for snow based on snow
             density or on liquid water and ice content in the snow pack. 
             Differing parameterizations are available. The result is an 
             effective thermal conductivity as various processes (conduction,
             advection, diffusion, radiative transfer) are parameterised. In the
             ground, differing mixing model are applied to the proprtions
             and thermo-physical properties of ground constituents.
}
\details{
Positive are to the right while left shifts are expressed as a
         negative number. All shifts are circular. Elements shifted off one
         end wrap around and are shifted onto the other end. This function
         mimicks the behaviour of SHIFT in IDL.
}
\examples{
rho.i <-  917  # [kg m-3]
rho.l <- 1000  # [kg m-3]
mat   <- data.frame(ice = (0:600)/1000, 
                    liq = rep(0.0,601))
dens  <- mat$ice*rho.i + mat$liq*rho.l

mat <- TermCond(mat, type.sno = "CALONNE", layers.sno = 1:601)

plot(dens, mat$kj, type = "l", col = "black", main = 
     "Comparison of parameterization for snow thermal conductivity", 
     xlab = "Density [kg m-3]", ylab = "Thermal conductivity [W m-1 K-1]")

mat <- TermCond(mat , type.sno = "YEN", layers.sno = 1:601)
lines(dens, mat$kj, col = "red")

mat <- TermCond(mat , type.sno = "COSENZA", layers.sno = 1:601)
lines(dens, mat$kj, col = "blue")

mat <- TermCond(mat , type.sno = "STURM", layers.sno = 1:601)
lines(dens, mat$kj, col = "green")

mat <- TermCond(mat , type.sno = "WILLIAMS", layers.sno = 1:601)
lines(dens, mat$kj, col = "orange")

mat <- ermCond(mat , type.sno = "JORDAN", layers.sno = 1:601)
lines(dens, mat$kj, col = "darkorchid2")

legend("topleft", inset=.05, seg.len = 2, lty = 1, lwd =1,
  	   c("CALONNE","YEN","COSENZA", "STURM", "WILLIAMS", "JORDAN"), 
  	   col = c("black", "red", "blue", "green", "orange", "darkorchid2"))

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
