% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trisol.R
\name{Trisol}
\alias{Trisol}
\title{Trisol}
\usage{
Trisol(a, b, c, d)
}
\arguments{
\item{a}{Lower diagonal term [Z], Note: a[,0] is zero}

\item{b}{Center diagonal term [Z]}

\item{c}{Upper diagonal term [Z], Note: c[.Z-1] is zero}

\item{d}{Right-hand side (on input) / solution (on output) [Z]}
}
\value{
Solution vector result. In case of heat conduction this is
       the temperature for the new time step.
}
\description{
Solves tri-diagonal matrix equations.
}
\details{
Solves matrix with Thomas algorithm, Z discretizations with depth.
}
\examples{
a <- c(0, 1, 1, 1)
b <- c(2, 2, 2, 2)
c <- c(1, 1, 1, 0)
d <- c(1, 1, 1, 1)
d <- Trisol(a, b, c, d)

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
