% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_data_import.R
\name{TSA_data_import}
\alias{TSA_data_import}
\title{Return data frame or a list of data frames with the daily aggregated data for a location
       or several locations.}
\usage{
TSA_data_import(con, inventory, time_b = "1950-01-01 00:00:00+00",
  time_e = "2050-01-01 00:00:00+00")
}
\arguments{
\item{con}{Connection to Permafrost data base}

\item{inventory}{A data frame with a column "name", a list or a character containing all the location names from which
the data should be imported}

\item{time_b}{Begin date for the import. Format: "1950-01-01 00:00:00+00"}

\item{time_e}{End date for the import. Format: "2050-01-01 00:00:00+00"}
}
\value{
List of data fames with the following fields: 
        Location name as loc_name;
        Observation height/depth as height [m];
        Mean daily GST as agg_avg [C];
        Number of daily observations as agg_cnt;
        Minimum daily GST as agg_min [C];
        Maximum daily GST as agg_max [C];
        Mean daily standard deviation as agg_sd;
        Date as time (in POSIXct and the following format: YYYY-MM-DD);
}
\description{
Give a data.frame, a list or a character of location-names and the function returns a data frame or a list 
             of data frames with daily aggregated data for each location.
}
\details{
The default settings are set as to import the whole dataset of a location. By adjusting the time_b and 
         time_e parameters also only single years or winter periods can be imported.
}
\examples{
con <- dbpf_con()
BC_SO01_01 <- TSA_data_import(con, "BC16-SO01_01")

}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
