% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_plot.R
\name{TSA_plot}
\alias{TSA_plot}
\title{Plot of the time series of a location}
\usage{
TSA_plot(obs_stat, out.path, v1 = 0.1, v2 = 0.3, lengthsnow = 5,
  MDr.sd = 0.3, v = 0.25, tperc = 40, sdoutlier = 3, sdsteep = 0.75,
  temp = 0.2, slopesd = 0.1, tempsd = 0.01, lengthzc = 2)
}
\arguments{
\item{obs_stat}{one dataframe of one location or a list of dataframes of multiple locations.
A dataframe needs at least the following columns:
Location name as loc_name;
Mean daily GST as agg_avg;
Minimum daily GST as agg_min;
Maximum daily GST as agg_max;
Daily sd as agg_sd;
Date as time (in POSIXct and the following format: YYYY-MM-DD)}

\item{out.path}{path to directory where the plot should be saved. e.g: "~/Desktop/"}

\item{v1}{max. daily standard deviation indicating snow for POSITIVE Ground Surface Temperature. 
Default is 0.1.}

\item{v2}{max. daily standard deviation indicating snow for NEGATIVE Ground Surface Temperature. 
Default is 0.3.}

\item{lengthsnow}{number of days that the snow cover should at least have to be selected as a valid snow cover.
Default is 5 days}

\item{MDr.sd}{threshold for the mean daily standard deviation for a snow period.}

\item{v}{threshold zero curtain (if dailymax & dailymin is within +/- v, this day is a zero curtain day)
Default: 0.25}

\item{tperc}{percentage of the 10-percent-quantile of the winter temperatures that the 
warming period should at least reach. Default is 40 percent.}

\item{sdoutlier}{The number implies how many multiples of the standard deviation of the slope are taken to define the threshold.
The threshold is used to detect slope outliers.}

\item{sdsteep}{The number implies what fraction of the standard deviation of the slope 
is taken to select the steepest part of a warming period.}

\item{temp}{the temperature boundary within the mean agg_avg of a zero curtain period has to be. 
Default: 0.2 degrees Celsius}

\item{slopesd}{Fraction of the standard deviation of the slope used as threshold for the slope. Default: 0.1}

\item{tempsd}{Fraction of the mean daily standard deviation used as threshold for the daily sd. Default: 0.01}

\item{lengthzc}{Length that a zero curtain period should at least have. Default: 2 days.}
}
\value{
Pdf-file saved in a own defined directory. Plot contains:
        min-, max-, mean- daily temperatures as well as snow cover, RD, warming periods, zero curtains
}
\description{
Give one data.frame of a location or a list of data.frames of multiple 
             locations and the function returns a plot containing the minimum, maximum, mean daily GST
             as well as the snow cover, RD, warming periods and zero curtains. The output is a
             pdf-file, saved in a own defined output directory
}
\details{

}
\examples{
con <- dbpf_con()
BC_SO01_01 <- TSA_data_import(con, "BC16-SO01_01")
TSA_snow_cover(BC_SO01_01, "~/Desktop/", v1 = 0.1, v2 = 0.3, lengthsnow = 5, MDr.sd = 0.3, v= 0.25, tperc = 40, sdoutlier = 3, sdsteep = 0.75, temp= 0.2, slopesd= 0.1, tempsd= 0.01, lengthzc= 2)

}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
