% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_RD.R
\name{TSA_RD}
\alias{TSA_RD}
\title{Detect RD date using the theory and function by Schmid et al. (2012)}
\usage{
TSA_RD(obs_stat, v = 0.25, v1 = 0.1, v2 = 0.3, lenghtsnow = 5,
  MDr.sd = 0.3, snow.period = NULL)
}
\arguments{
\item{obs_stat}{One dataframe of one location or a list of dataframes of multiple locations.
A dataframe needs at least the following columns:
Location name as loc_name;
Mean daily GST as agg_avg;
Minimum daily GST as agg_min;
Maximum daily GST as agg_max;
Daily sd as agg_sd;
Date as time (in POSIXct and the following format: YYYY-MM-DD)}

\item{v}{Threshold zero curtain (if dailymax & dailymin is within +/- v, this day is a zero curtain day)
Default: 0.25}

\item{v1}{Max. daily standard deviation indicating snow for POSITIVE Ground Surface Temperature. 
Default is 0.1.}

\item{v2}{Max. daily standard deviation indicating snow for NEGATIVE Ground Surface Temperature. 
Default is 0.3.}

\item{MDr.sd}{Threshold for the mean daily standard deviation for a snow period.}

\item{snow.period}{In case the snow period is already available for one location, it can be inserted here.
The data frame of the now period needs to have the following columns: 
loc_name, Snow.Start, Snow.End, MDr}

\item{lengthsnow}{Number of days that the snow cover should at least have to be selected as a valid snow cover.
Default is 5 days.}
}
\value{
Dataframe with fields: loc_name; Snow.Start; Snow.End; MDr
}
\description{
Give one data.frame of a location or a list of data.frames of multiple 
             locations and the function returns the Ripening Date for the snow periods.
}
\details{
It is also possible to adjust the parameters for the TSA_snow_cover function in this function, 
         as the warming periods are calculated for each snow cover period. 
         Therefore, the TSA_snow_cover is run as well in this function.
}
\examples{
con <- dbpf_con()
BC_SO01_01 <- TSA_data_import(con, "BC16-SO01_01")
RD_table <- TSA_RD(BC_SO01_01, v= 0.25, m1 = 3, m2 = 0.5, v1 = 0.1, v2 = 0.3, lenghtsnow = 5, MDr.sd = 0.3, snow.period=NULL)

}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
