% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_row.R
\name{TSA_row}
\alias{TSA_row}
\title{Return an empty data frame containing loc_name, Snow.Start and Snow.End}
\usage{
TSA_row(location_name)
}
\arguments{
\item{obs_stat}{data frame with a index-column}
}
\value{
a dataframe with loc_name, Snow.Start=NA, Snow.End=NA
}
\description{
Creates an empty data frame containing loc_name, Snow.Start and Snow.End.
             The Loc.name is inserted.
}
\details{

}
\examples{
table <- TSA_row(location_name)


}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
\keyword{internal}
