% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_snow_cover.R
\name{TSA_snow_cover}
\alias{TSA_snow_cover}
\title{Find start and end date of snow cover}
\usage{
TSA_snow_cover(obs_stat, v1 = 0.1, v2 = 0.3, lengthsnow = 5,
  MDr.sd = 0.3)
}
\arguments{
\item{obs_stat}{One dataframe of one location or a list of dataframes of multiple locations.
A dataframe needs at least the following columns:
Location name as loc_name;
Mean daily GST as agg_avg;
Minimum daily GST as agg_min;
Maximum daily GST as agg_max;
Daily sd as agg_sd;
Date as time (in POSIXct and the following format: YYYY-MM-DD)}

\item{v1}{Max. daily standard deviation indicating snow for POSITIVE Ground Surface Temperature. 
Default is 0.1.}

\item{v2}{Max. daily standard deviation indicating snow for NEGATIVE Ground Surface Temperature. 
Default is 0.3.}

\item{lengthsnow}{Number of days that the snow cover should at least have to be selected as a valid snow cover.
Default is 5 days}

\item{MDr.sd}{Threshold for the mean daily standard deviation for a snow period.}
}
\value{
Dataframe with fields: loc_name; Snow.Start; Snow.End; MDr
}
\description{
Give one data.frame of a location or a list of data.frames of multiple 
             locations and the function returns the start and end date as well as  
             the MDr in a output table
}
\details{
At locations with perennial snow cover, the snow cover is splitted into water years periods.
}
\examples{
con <- dbpf_con()
BC_SO01_01 <- TSA_data_import(con, "BC16-SO01_01")
Snow_table <- TSA_snow_cover(BC_SO01_01, v1 = 0.1, v2 = 0.3, lengthsnow = 5, MDr.sd = 0.3)

}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
