% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_subsetting.R
\name{TSA_subsetting}
\alias{TSA_subsetting}
\title{Return list with subset dataframes}
\usage{
TSA_subsetting(obs_stat)
}
\arguments{
\item{obs_stat}{data frame with a index-column}
}
\value{
a list containing all the subset-dataframes
}
\description{
Give one data.frame of an index-column containing unique values. It returns a list of 
             the subset-dataframes
}
\details{

}
\examples{
snow_periodlist <- TSA_subsetting(obs_stat)


}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
\keyword{internal}
