% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_warming_periods.R
\name{TSA_warming_periods}
\alias{TSA_warming_periods}
\title{Detect major warming periods during a snow cover period}
\usage{
TSA_warming_periods(obs_stat, tperc = 40, sdoutlier = 3, sdsteep = 0.75,
  v1 = 0.1, v2 = 0.3, lengthsnow = 5, MDr.sd = 0.3,
  snow.period = NULL)
}
\arguments{
\item{obs_stat}{One dataframe of one location or a list of dataframes of multiple locations
A dataframe needs at least the following columns:
Location name as loc_name;
Mean daily GST as agg_avg;
Minimum daily GST as agg_min;
Maximum daily GST as agg_max;
Daily sd as agg_sd; 
Date as time (in POSIXct and the following format: YYYY-MM-DD)}

\item{tperc}{Percentage of the 10-percent-quantile of the winter temperatures that the 
warming period should at least reach. Default is 40 percent.}

\item{sdoutlier}{The number implies how many multiples of the standard deviation of the slope are taken to define a threshold.
The threshold is used to detect slope outliers.}

\item{sdsteep}{The number implies what fraction of the standard deviation of the slope 
is taken to select the steepest part of a warming period.}

\item{v1}{Max. daily standard deviation indicating snow for POSITIVE Ground Surface Temperature. 
Default is 0.1.}

\item{v2}{Max. daily standard deviation indicating snow for NEGATIVE Ground Surface Temperature. 
Default is 0.3.}

\item{lengthsnow}{Number of days that the snow cover should at least have to be selected as a valid snow cover.
Default is 5 days.}

\item{MDr.sd}{Threshold for the mean daily standard deviation for a snow period.}

\item{snow.period}{In case the snow period is already available for one location, it can be inserted here.
The data frame of the now period needs to have the following columns: 
loc_name, Snow.Start, Snow.End, MDr}
}
\value{
Datafame with the following fields: 
        loc_name: location name;
        Snow.Start: Start date of the analyzed snow period; 
        Snow.End: End date of the analyzed snow period;
        Warming.Period.Start: start date of the warming period;
        RD: indicates if the warming period is assumed to be the Ripening date; 
        Temp.diff: Temperature of the steepest part of the warming period [C];
        Mean.Slope: Mean slope of the steepest part of the warming period; 
        Temp.To.Zero: How much below/above zero is the end of the steepest part of the warming period [C];
}
\description{
Give one data.frame of a location or a list of data.frames of multiple 
             locations and the function returns a table with the start dates of the major warming periods  
             during a winter as well as quality values.
}
\details{
It is also possible to adjust the parameters for the TSA_snow_cover function in this function, 
         as the warming periods are calculated for each snow cover period. 
         Therefore, the TSA_snow_cover is run as well in this function.
}
\examples{
con <- dbpf_con()
BC_SO01_01 <- TSA_data_import(con, "BC16-SO01_01")
WarmingPeriodsTable <- TSA_warming_periods(BC_SO01_01,tperc = 40, sdoutlier = 3, sdsteep = 0.75, v1 = 0.1, v2 = 0.3, lengthsnow = 5, MDr.sd = 0.3, snow.period=NULL)

}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
