% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSA_zero_curtain.R
\name{TSA_zero_curtain}
\alias{TSA_zero_curtain}
\title{Detect zero curtain periods during a snow cover period}
\usage{
TSA_zero_curtain(obs_stat, temp = 0.2, slopesd = 0.1, tempsd = 0.01,
  lengthzc = 2, v1 = 0.1, v2 = 0.3, lengthsnow = 5, MDr.sd = 0.3,
  snow.period = NULL)
}
\arguments{
\item{obs_stat}{one dataframe of one location of a list of dataframes of multiple locations
A dataframe needs at least the following columns:
Location name as loc_name;
Mean daily GST as agg_avg;
Maximum daily GST as agg_max;
Daily sd as agg_sd (not necessarily needed);
Date as time (in POSIXct and the following format: YYYY-MM-DD)}

\item{temp}{The temperature boundary within the mean agg_avg of a zero curtain period has to be. 
Default: 0.2 degrees Celsius}

\item{slopesd}{Fraction of the standard deviation of the slope used as threshold for the slope. Default: 0.1}

\item{tempsd}{Fraction of the mean daily standard deviation used as threshold for the daily sd. Default: 0.01}

\item{lengthzc}{Length that a zero curtain period should at least have. Default: 2 days.}

\item{v1}{Max. daily standard deviation indicating snow for POSITIVE Ground Surface Temperature. 
Default is 0.1.}

\item{v2}{Max. daily standard deviation indicating snow for NEGATIVE Ground Surface Temperature. 
Default is 0.3.}

\item{lengthsnow}{Number of days that the snow cover should at least have to be selected as a valid snow cover.
Default is 5 days}

\item{MDr.sd}{Threshold for the mean daily standard deviation for a snow period.}

\item{snow.period}{In case the snow period is already available for one location, it can be inserted here.
The data frame of the now period needs to have the following columns: 
loc_name, Snow.Start, Snow.End, MDr}
}
\value{
Data fame with the following fields: 
        loc_name: location name;
        Snow.Start: Start date of the analyzed snow period; 
        Snow.End: End date of the analyzed snow period;
        Zero.Curtain.Start: Start date of a zero curtain period;
        Zero.Curtain.End: End date of a zero curtain period;
        Median.Temperature: Median temperature of a zero curtain period [C];
        sd.Temperature: Standard deviation of the temperature of a zero curtain period;
}
\description{
Give one data.frame of a location or a list of data.frames of multiple 
             locations and the function returns a table with the start and end dates of zero curtain
             periods during snow cover periods.
}
\details{
It is also possible to adjust the parameters for the TSA_snow_cover function in this function, 
         as the warming periods are calculated for each snow cover period. 
         Therefore, the TSA_snow_cover is run as well in this function.
}
\examples{
con <- dbpf_con()
BC_SO01_01 <- TSA_data_import(con, "BC16-SO01_01")
ZeroCurtainTable <- TSA_zero_curtain(BC_SO01_01, temp= 0.2, slopesd= 0.1, tempsd= 0.01, lengthzc=2, v1 = 0.1, v2 = 0.3, lengthsnow = 5, MDr.sd = 0.3, snow.period=NULL)

}
\author{
Thomas Knecht <t.knecht@hotmail.com>
}
