% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfrozen.R
\name{Unfrozen}
\alias{Unfrozen}
\title{Unfozen water in soil}
\usage{
Unfrozen(mat, unfrozen.type = "INTERVAL", unfrozen.par = 0.5)
}
\arguments{
\item{mat}{Data frame with relevant ground properties.}

\item{unfrozen.type}{Character string indicating the type of function to be
                     used; the standard is "INTERVAL". The full options are: 
                     "DALLAMICO": Dall’Amico, M., Endrizzi, S., Gruber, S., 
                     & Rigon, R. (2011). A robust and energy-conserving 
                     model of freezing variably-saturated soil. The 
                     Cryosphere, 5(2), 469–484. doi:10.5194/tc-5-469-2011
                     Typical values of van Genuchten parameters are found 
                     in Table 2 of Gubler, S., Endrizzi, S., Gruber, S., 
                     & Purves, R. S. (2013). Sensitivities and uncertainties 
                     of modeled ground temperatures in mountain environments. 
                     Geoscientific Model Development, 6(4), 1319–1336. 
                     doi:10.5194/gmd-6-1319-2013

                     "MOTTAGHY": Mottaghy, D., & Rath, V. (2006). Latent heat 
                     effects in subsurface heat transport modelling and their
                     impact on palaeotemperature reconstructions. Geophysical 
                     Journal International, 164(1), 236-245.

                     "INTERVAL": Phase change takes place in an interval of
                     specified with (unfrozen.par) below 0C and at a 
                     constant rate.}

\item{unfrozen.par}{Parameter set for the chosen unfrozen water function.

                    "DALLAMICO": unfrozen.par[1]: van Genuchten alpha [mm-1], 
                    unfrozen.par[2]: van Genuchten n [-], unfrozen.par[3]: 
                    residual water content [m3/m3]. The saturated water 
                    content is given by the input (mat$wat) and can thus 
                    vary with depth. 

                    "MOTTAGHY": unfrozen.par[1]: width of freezing 
                    interval [K], unfrozen.par[2]: omega.  

                    "INTERVAL" unfrozen.par[1]: width of freezing 
                    interval [K].}
}
\value{
Returns the input data frame with three columns updated: liq [m3/m3]
        (liquid water content, relative to soil volume), ice [m3/m3] (ice
        content, relative to soil volume), and dice [m3/C] (change in ice
        content, relative to soil volume per degree Celsius). If these columns
        do not exist in the input data frame, they are created.
}
\description{
Computes the proportion of unfozen water and ice in soil as 
             well as the temperature derivative of liquid water content.
}
\details{
This function applies invariant freezing parameters with depth and
         assumes saturated conditions without movement of water.
}
\examples{
mat <- data.frame(Tj  = (-500:200)/100, 
                  wat = rep(0.5,701))

#Mottahgy and Rath (2006) example
mat <- Unfrozen(mat, unfrozen.type = "MOTTAGHY", unfrozen.par = c(0, 0.5))
plot(mat$Tj, mat$dice, type="l", lty = 1, col = "black", ylim =c(0,2),
     xlab = "Temperature [C]", 
     ylab = "Liquid water content (solid), derivation (dashed)")
lines(mat$Tj, mat$liq, lty = 2, col = "black")

#interval function example
mat <- Unfrozen(mat, unfrozen.type = "INTERVAL", unfrozen.par = 1)
lines(mat$Tj, mat$dice, lty = 1, col = "blue")
lines(mat$Tj, mat$liq,  lty = 2, col = "blue")

#Dall'Amico et al. (2011) example, assuming saturated conditions.
mat <- Unfrozen(mat, unfrozen.type = "DALLAMICO", 
                unfrozen.par = c(0.001, 1.4, 0.05))
lines(mat$Tj, mat$dice, lty = 1, col = "red")
lines(mat$Tj, mat$liq,  lty = 2, col = "red")

}
\author{
Stephan Gruber <stephan.gruber@carleton.ca>
}
